Add ability to supply file paths as arugments
This patch makes it possible to supply paths to files or directories as
trailing arguments to the CLI.

If a supplied path is a directory, all files in that directory will be
added recursively.

New error indicating that one or more of the provided paths are invalid
was added (and included in the list of handled errors).

Update usage banner.