Implement draft version of core functionality
This patch adds the core of what the application should be doing. The
CommandLineOutput class will surely need to be refactored and
a dedicated filtering class will most likely be needed.

While still buggy, this patch makes it possible to try out the binary
and have it do something that resembles the end result.