Make argument parser not find dotfiles by default
Right now, whether paths are provided as arguments or not, all files are
found recursively, including dotfiles. The default should be that
they're not found, while finding them should be possible if a flag is
passed to the application.

This patch doesn't add that flag, but it does set the default.