Catch more unreadable file errors
While ArgumentError will prevent the program from crashing when there's
a sudden unreadable character, a different error (Errno::ENOENT) is
raised when a non-text file is encountered, such as an image file.

Expand the begin-rescue-end clause so that everything inside the
select-block is surrounded and add Errno:ENOENT to the list of
exceptions that get caught.